Moved getUrlStringfromUrlObject function to utils file in curl and swift
