+++ codegens/curl/lib/util.js
@@ -110,5 +110,42 @@ module.exports = {
       }
     }
     return result;
+  },
+  /**
+ *
+ * @param {*} urlObject The request sdk request.url object
+ * @returns {String} The final string after parsing all the parameters of the url including 
+ * protocol, auth, host, port, path, query, hash
+ * This will be used because the url.toString() method returned the URL with non encoded query string
+ * and hence a manual call is made to getQueryString() method with encode option set as true.
+ */
+getUrlStringfromUrlObject: function (urlObject) {
+  var url = '';
+  if (urlObject.protocol) {
+    url += (urlObject.protocol.endsWith('://') ? urlObject.protocol : urlObject.protocol + '://');
   }
+  if (urlObject.auth && urlObject.auth.user) {
+    url = url + ((urlObject.auth.password) ?
+      // ==> username:password@
+      urlObject.auth.user + ':' + urlObject.auth.password : urlObject.auth.user) + '@';
+  }
+  if (urlObject.host) {
+    url += urlObject.getHost();
+  }
+  if (urlObject.port) {
+    url += ':' + urlObject.port.toString();
+  }
+  if (urlObject.path) {
+    url += urlObject.getPath();
+  }
+  if (urlObject.query && urlObject.query.count()) {
+    let queryString = urlObject.getQueryString({ ignoreDisabled: true, encode: true });
+    queryString && (url += '?' + queryString);
+  }
+  if (urlObject.hash) {
+    url += '#' + urlObject.hash;
+  }
+
+  return url;
+}
 };
